\name{RadOnc-package}
\alias{RadOnc-package}
\alias{RadOnc}
\docType{package}
\title{
Analytical Tools for Radiation Oncology
}
\description{
This package is designed for the import, analysis, and visualization of dosimetric and volumetric data in Radiation Oncology.  The tools herein enable import of DVH information from Varian's Aria/Eclipse platform and 3D structural representations from DICOM-RT files.  These tools also enable subsequent visualization and statistical analysis of this data.
}
\details{
\tabular{ll}{
Package: \tab RadOnc\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2013-10-21\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Reid F. Thompson <reid.thompson@gmail.com>
}
\keyword{ package }
\seealso{
\code{\link[RadOnc:DVH-class]{DVH}}, \code{\link[RadOnc:DVH.list-class]{DVH.list}}, \code{\link[RadOnc:structure3D-class]{structure3D}}, \code{\link[RadOnc:structure.list-class]{structure.list}}
}
