\name{RadOnc-package}
\alias{RadOnc-package}
\alias{RadOnc}
\docType{package}
\title{
\packageTitle{RadOnc}
}
\description{
\packageDescription{RadOnc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RadOnc}
\packageIndices{RadOnc}
}
\author{
\packageAuthor{RadOnc}

Maintainer: \packageMaintainer{RadOnc}
}
\references{
Thompson, R.F. (2014)
RadOnc: An R Package for Analysis of Dose-Volume Histogram and Three-Dimensional Structural Data.
\emph{J Radiat Oncol Inform} \bold{6}(1):98-110.
}
\keyword{ package }
\seealso{
\code{\link[RadOnc:DVH-class]{DVH}}, \code{\link[RadOnc:DVH.list-class]{DVH.list}}, \code{\link[RadOnc:structure3D-class]{structure3D}}, \code{\link[RadOnc:structure.list-class]{structure.list}}, \code{\link[RadOnc:RTdata-class]{RTdata}}
}
\examples{
## See sub-documentation for detailed examples of important package functions ##
}
